Refactor : sprint3 피드백에 대한 수정

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicReadStatusService.java
@@ -5,8 +5,8 @@ import com.sprint.mission.discodeit.repository.ChannelRepository;
 import com.sprint.mission.discodeit.repository.ReadStatusRepository;
 import com.sprint.mission.discodeit.repository.UserRepository;
 import com.sprint.mission.discodeit.service.ReadStatusService;
-import com.sprint.mission.discodeit.service.dto.readstatus.ReadStatusCreateParam;
-import com.sprint.mission.discodeit.service.dto.readstatus.ReadStatusUpdateParam;
+import com.sprint.mission.discodeit.service.dto.readstatus.ReadStatusCreateRequest;
+import com.sprint.mission.discodeit.service.dto.readstatus.ReadStatusUpdateRequest;
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.UUID;
@@ -21,16 +21,17 @@ public class BasicReadStatusService implements ReadStatusService {
     private final ReadStatusRepository readStatusRepository;
 
     @Override
-    public void create(ReadStatusCreateParam param) {
+    public void create(ReadStatusCreateRequest param) {
         param.userIds().forEach(userId -> {
             if (!userRepository.existsById(userId)) {
                 throw new IllegalArgumentException(userId + " 에 해당하는 User을 찾을 수 없음");
             }
-            readStatusRepository.findAllByUserId(userId).forEach(readStatus -> {
-                if (readStatus.getChannelId().equals(param.channelId())) {
-                    throw new IllegalArgumentException("중복된 객체 존재");
-                }
-            });
+            boolean isDuplicate = readStatusRepository.findAllByUserId(userId)
+                    .stream().anyMatch(readStatus -> readStatus.getChannelId().equals(param.channelId()));
+            if (isDuplicate) {
+                throw new IllegalArgumentException("중복된 객체 존재");
+            }
+
             readStatusRepository.save(new ReadStatus(userId, param.channelId()));
         });
     }
@@ -50,11 +51,12 @@ public class BasicReadStatusService implements ReadStatusService {
     }
 
     @Override
-    public List<UUID> findAllUserByChannelId(UUID channelId) { // channelId와 동일한 channelId에 포함된 User의 UUID 모음
+    public List<UUID> findAllUserByChannelId(UUID channelId) {
         if (!channelRepository.existsById(channelId)) {
             throw new IllegalArgumentException(channelId + " 에 해당하는 Channel를 찾을 수 없음.");
         }
-        return readStatusRepository.findAllUserByChannelId(channelId);
+        return readStatusRepository.findAllByChannelId(channelId)
+                .stream().map(ReadStatus::getUserId).toList();
     }
 
     @Override
@@ -62,17 +64,22 @@ public class BasicReadStatusService implements ReadStatusService {
         if (!channelRepository.existsById(channelId)) {
             throw new IllegalArgumentException(channelId + " 에 해당하는 Channel를 찾을 수 없음.");
         }
-        return readStatusRepository.findAllByChannelId(channelId);
+        return readStatusRepository.findAllByChannelId(channelId)
+                .stream().map(ReadStatus::getId).toList();
     }
 
     @Override
-    public void update(ReadStatusUpdateParam param) {
+    public void update(ReadStatusUpdateRequest param) {
         ReadStatus readStatus = find(param.id());
+        readStatus.update();
         readStatusRepository.save(readStatus);
     }
 
     @Override
     public void delete(UUID id) {
+        if (!readStatusRepository.existsById(id)) {
+            throw new IllegalArgumentException(id + " 에 해당하는 ReadStatus를 찾을 수 없음.");
+        }
         readStatusRepository.delete(id);
     }
 }
